feature: first release (#37)
* feat: 实现skipWriter

* changes: need fix CompetitiveImpactAccumulator

* changes: skipReader

* feat: 完善skipreader的ReadSkipData函数

* feat: 优化skipReader的实现

* changes: 增加skipReader的interface的context参数

* changes: 增加context参数

* changes: 增加context参数

* changes: add context

* changes: 修复部分索引问题

* changes: 修复indexwriter的运行失败问题

* fix: 修复segment数量问题
todo: writeIndexOutput写入数据异常修复

* fix: 修复scf写入异常

* fix: 修复CompoundDirectory

* feat: 设置context

* fix: 修复reader读取索引失败

* fix: 修复NORMS_SEG_EXTENSION写入错误

* changes: 简单调整下代码

* changes: 实现freqproxfields部分代码

* changes: 修复bug

* fix: 修复buffer获取问题

* docs: 更新文档

+++ codecs/simpletext/storedfieldsreader.go
@@ -204,7 +204,7 @@ func (s *StoredFieldsReader) readField(
 		if err != nil {
 			return err
 		}
-		return visitor.Int64Field(fieldInfo, int64(num))
+		return visitor.Int64Field(fieldInfo, num)
 	case bytes.Equal(STORED_FIELD_TYPE_FLOAT, dataType):
 		num, err := strconv.ParseFloat(string(value), 32)
 		if err != nil {
@@ -216,7 +216,7 @@ func (s *StoredFieldsReader) readField(
 		if err != nil {
 			return err
 		}
-		return visitor.Float64Field(fieldInfo, float64(num))
+		return visitor.Float64Field(fieldInfo, num)
 	default:
 		return errors.New("unknown field type")
 	}
